refactor: move to separate component and move button to bottom of screen
